pairing with julia about function extraction/componentization
