+++ src/componentsNext/NotificationMenu/config.ts
@@ -1,3 +1,5 @@
+import { LintSchemaType } from "~/api/namespaceLinting/schema";
+import { NotificationItemType } from "./NotificationItem";
 import { SquareAsterisk } from "lucide-react";
 import { pages } from "~/util/router/pages";
 import { useNamespace } from "~/util/store/namespace";
@@ -19,4 +21,27 @@ export const useNotificationConfig = () => {
       }),
     },
   } as const;
+};
+
+export const useGroupNotifications = (data: LintSchemaType | undefined) => {
+  const notificationConfig = useNotificationConfig();
+  const notificationTypes = Object.entries(notificationConfig ?? {});
+  return notificationTypes
+    .map(([notificationType, notificationConfig]) => {
+      const matchingNotifications = data?.issues.filter(
+        (issue) => notificationType === issue.type
+      );
+
+      if (!matchingNotifications || matchingNotifications.length === 0) {
+        return null;
+      }
+
+      const { icon, description, href } = notificationConfig;
+      return {
+        icon,
+        description: description(matchingNotifications.length),
+        href,
+      };
+    })
+    .filter((item) => item !== null) as NotificationItemType[];
 };
