+++ src/componentsNext/NotificationMenu/index.tsx
@@ -1,64 +1,28 @@
-import { Check, LucideIcon, SquareAsterisk } from "lucide-react";
 import {
   Notification,
-  NotificationClose,
   NotificationLoading,
   NotificationMenuSeparator,
   NotificationMessage,
   NotificationTitle,
 } from "~/design/Notification/";
 
-import { Link } from "react-router-dom";
+import { Check } from "lucide-react";
+import { Fragment } from "react";
+import { NotificationItem } from "./NotificationItem";
 import { twMergeClsx } from "~/util/helpers";
+import { useGroupNotifications } from "./config";
 import { useNamespaceLinting } from "~/api/namespaceLinting/query/useNamespaceLinting";
-import { useNotificationConfig } from "./config";
 import { useTranslation } from "react-i18next";
 
-function NotificationItem({
-  href,
-  description,
-  icon: Icon,
-}: {
-  href: string;
-  description: string;
-  icon: LucideIcon;
-}) {
-  return (
-    <div>
-      <div>{href}</div>
-      <div>{description}</div>
-      <Icon />
-    </div>
-  );
-}
-
-interface NotificationMenuProps {
+type NotificationMenuProps = {
   className?: string;
-}
+};
 
 const NotificationMenu: React.FC<NotificationMenuProps> = ({ className }) => {
   const { t } = useTranslation();
   const { data, isLoading } = useNamespaceLinting();
-  const notificationConfig = useNotificationConfig();
   const showIndicator = !!data?.issues.length;
-
-  const notificationTypes = Object.entries(notificationConfig ?? {});
-
-  const notificationItems = notificationTypes.map(
-    ([notificationType, notificationConfig], index, srcArr) => {
-      const matchingNotifications = data?.issues.filter(
-        (issue) => notificationType === issue.type
-      );
-      if (
-        matchingNotifications == undefined ||
-        matchingNotifications.length <= 0
-      ) {
-        return null;
-      }
-
-      let { icon, description, count, href } = srcArr[index];
-    }
-  );
+  const notificationItems = useGroupNotifications(data);
 
   return (
     <div className={twMergeClsx("self-end text-right", className)}>
@@ -72,18 +36,14 @@ const NotificationMenu: React.FC<NotificationMenuProps> = ({ className }) => {
             {t("components.notificationMenu.isLoading.description")}
           </NotificationLoading>
         )}
-        {showIndicator && notificationItems != null ? (
+        {showIndicator ? (
           notificationItems.map((item, index) => {
             const isLastListItem = index === notificationItems.length - 1;
             return (
-              <div key="">
-                <NotificationItem
-                  icon={item.icon}
-                  href={item.href}
-                  description={item.description}
-                />
+              <Fragment key={index}>
+                <NotificationItem {...item} />
                 {!isLastListItem && <NotificationMenuSeparator />}
-              </div>
+              </Fragment>
             );
           })
         ) : (
